added cd crud validations

+++ src/app/cd-list/cd-list.component.ts
@@ -3,12 +3,14 @@ import { MatDialog } from '@angular/material/dialog';
 import { CdService } from '../services/cd.service';
 import { CdAddComponent } from '../cd-add/cd-add.component';
 import { Router } from '@angular/router';
+import { ContactService } from '../services/contact.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cd-list',
   templateUrl: './cd-list.component.html',
   styleUrls: ['./cd-list.component.css'],
-  providers: [CdService]
+  providers: [CdService, ContactService]
 })
 export class CdListComponent implements OnInit {
   name!: string;
@@ -18,18 +20,61 @@ export class CdListComponent implements OnInit {
   constructor(
     public dialog: MatDialog,
     private cdService: CdService,
+    private contactService: ContactService,
     private router: Router
   ) { }
 
   async ngOnInit(): Promise<void> {
     const cd = await this.cdService.all();
+    this.getNames(cd);
     this.dataSource.push(cd);
     this.dataSource = [...this.dataSource];
   }
 
   async deleteCd(id: any): Promise<void> {
-    await this.cdService.delete(id);
-    this.reload('/cd');
+    const result = await Swal.fire({
+      title: 'Atenção!',
+      text: 'Você tem certeza que deseja deletar esse registro?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Continuar'
+    });
+    if (result.isConfirmed) {
+      await this.cdService.delete(id);
+      this.reload('/cd');
+      const Toast = Swal.mixin({
+        toast: true,
+        position: 'top-end',
+        showConfirmButton: false,
+        timer: 3000,
+        timerProgressBar: true,
+        didOpen: (toast) => {
+          toast.addEventListener('mouseenter', Swal.stopTimer);
+          toast.addEventListener('mouseleave', Swal.resumeTimer);
+        }
+      });
+      Toast.fire({
+        icon: 'success',
+        title: 'registro deletado com sucesso.'
+      });
+    }
+  }
+
+  async getNames(cd: any): Promise<void> {
+    if (cd) {
+      for (let i = 0; i < cd.length; i++) {
+        if (cd[i].contact_id) {
+          const contact = await this.contactService.find(cd[i].contact_id);
+          cd[i].contact_id = contact.name;
+        } else {
+          cd[i].contact_id = '';
+        }
+      }
+    }
+    return;
   }
 
   cdAdd(cd: any): void {
